Process invite on invite creation if user already exist + refacto
