+++ functions/src/triggers/userInvite.ts
@@ -4,6 +4,7 @@ import send from '../email/send'
 import userInvited from '../email/templates/userInvited'
 import {isEmpty} from 'lodash'
 import {Response} from 'node-fetch'
+import {arrayUnion, serverTimestamp} from '../helpers/firebaseInit'
 
 // userInviteCreated is called when an user as been invited to a project (to have edit rights).
 // An email is sent to him with the invitation link.
@@ -34,18 +35,94 @@ export const userInviteCreated = functions.firestore
             html: userInvited(
                 `${data.originUserName} invited you to become a member of the event "${data.projectName}"`,
                 `Hey ${data.destinationUserInfo},<br/><br/> I invited you to become a member of the OpenFeedback event <b>${data.projectName}</b>. Click on the button below to accept my invitation.`,
-                'Accept the invitation',
+                'View the event',
                 `${app.url}/admin/?inviteId=${inviteId}`
             )
         })
         if (sendResult instanceof Response && sendResult.ok) {
-            return admin
+            const promises: Promise<any>[] = []
+            promises.push(admin
                 .firestore()
                 .collection('projects-invites')
                 .doc(inviteId)
                 .update({
                     status: 'emailSent'
-                })
+                }))
+            promises.push(
+                admin
+                    .firestore()
+                    .collection('users')
+                    .where('email', '==', data.destinationUserInfo)
+                    .limit(1)
+                    .get()
+                    .then(querySnapshot => {
+                        if(querySnapshot.empty) {
+                            return Promise.resolve('no user matched')
+                        }
+                        let promise: Promise<any> = Promise.resolve('forEach not implemented or firebase issue')
+                        querySnapshot.forEach(userSnapshot => {
+                            promise = checkPendingInviteAndProcessThem(userSnapshot.data() as admin.auth.UserRecord)
+                        })
+                        return promise
+                    })
+            )
+
+            return Promise.all(promises)
         }
-        return Promise.reject('Function failed due to ??')
+
+        return Promise.reject('Function failed due to not successful email send')
     })
+
+// From a given user, it check if the user has pending invite for him, if so process it, add it to the linked project
+// and complete the invite after
+export const checkPendingInviteAndProcessThem = (user: admin.auth.UserRecord) => {
+    const userId = user.uid
+    const userInfo = user.email || user.phoneNumber
+
+    return admin
+        .firestore()
+        .collection('projects-invites')
+        .where('destinationUserInfo', '==', userInfo)
+        .where('status', '==', 'emailSent')
+        .get()
+        .then(querySnapshot => {
+            const promises: Promise<any>[] = []
+            querySnapshot.forEach(snapshot => {
+                const data = snapshot.data()
+                promises.push(new Promise((resolve, reject) => {
+                    admin
+                        .firestore()
+                        .collection('projects')
+                        .doc(data.projectId)
+                        .update({
+                            members: arrayUnion(userId),
+                            updatedAt: serverTimestamp()
+                        })
+                        .catch((error) => {
+                            return admin
+                                .firestore()
+                                .collection('projects-invites')
+                                .doc(snapshot.id)
+                                .update({
+                                    updatedAt: serverTimestamp(),
+                                    status: 'error',
+                                    error: JSON.stringify(error)
+                                })
+                        })
+                        .then(() => {
+                            return admin
+                                .firestore()
+                                .collection('projects-invites')
+                                .doc(snapshot.id)
+                                .update({
+                                    updatedAt: serverTimestamp(),
+                                    status: 'completed'
+                                })
+                        })
+                        .then((result) => resolve(result))
+                        .catch((error) => reject(error))
+                }))
+            })
+            return Promise.all(promises)
+        })
+}
